Add an AI challenge to the `fetch` day
I haven't made the `fullstack-meta` change yet since I'm not sure of the
best approach. This could either _replace_ the weather challenge or be a
full new challenge.

There aren't so many challenges on this day, and the weather one is
actually quite good, so I'm wondering if adding a new challenge might
even be best. There's no `rake` so why not?

resolve https://github.com/lewagon/fullstack-solutions/issues/335